requirements.txt
python-whois
requests